// frontend/src/models/user.models.ts
export interface User {
  id: number;
  name: string;
  email: string;
  role: 'User' | 'Admin';
  createdAt: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}